Move all envoyconfigrevision reconcile logic to the reconcilers package
