+++ pkg/reconcilers/marin3r/envoyconfigrevision/finalize.go
@@ -1,4 +1,4 @@
-package controllers
+package reconcilers
 
 import (
 	marin3rv1alpha1 "github.com/3scale/marin3r/apis/marin3r/v1alpha1"
