+++ pkg/reconcilers/marin3r/envoyconfigrevision/status.go
@@ -1,4 +1,4 @@
-package controllers
+package reconcilers
 
 import (
 	"fmt"
