+++ pkg/reconcilers/marin3r/envoyconfigrevision/status_test.go
@@ -1,4 +1,4 @@
-package controllers
+package reconcilers
 
 import (
 	"testing"
